Render song.html from a single document walk instead of per-field slots
SongDocumentRenderer.renderDocument walks doc.blocks in document order
and dispatches per block type (PBlock/RowBlock/LyricBlock/ScriptureBlock/
YoutubeBlock/GapBlock), so song.html collapses to one {{$song_content}}
placeholder instead of a fixed set of {{div:title}}/{{div:tune}}/etc.
slots that silently dropped any block type the template didn't know
about. The old lyrics-only render() is kept as renderLyrics() for the
bridge test's legacy-HTML comparison.

Scripture-reference rendering (previously duplicated as private methods
on SongViewActivity) is extracted into ScriptureReferenceRenderer, now
shared by SongFragment's in-document rendering and SongViewActivity's
copy/share text.
