+++ Alkitab/src/main/java/yuku/alkitab/songs/ScriptureReferenceRenderer.kt
@@ -1,0 +1,129 @@
+package yuku.alkitab.songs
+
+import yuku.alkitab.base.App
+import yuku.alkitab.base.util.AppLog
+import yuku.alkitab.base.util.OsisBookNames
+
+private const val TAG = "ScriptureReferenceRenderer"
+
+/**
+ * Converts scripture ref lines like `B1.C1.V1-B2.C2.V2; B3.C3.V3` into localized, readable text,
+ * optionally linked: `<a href='protocol:B1.C1.V1-B2.C2.V2'>Book 1 c1:v1-v2</a>; ...`. Localizes
+ * book names against the active Bible version. Shared by [SongViewActivity] (copy/share plain
+ * text) and [SongFragment] (in-document HTML rendering of a `ScriptureBlock`).
+ */
+object ScriptureReferenceRenderer {
+    /**
+     * @param protocol null to output plain text (no links); non-null to wrap each reference in an
+     * `<a href='protocol:osisId'>` link.
+     * @param line scripture ref(s) in OSIS
+     */
+    @JvmStatic
+    fun render(protocol: String?, line: String?): String {
+        if (line.isNullOrBlank()) return ""
+
+        val sb = StringBuilder()
+
+        val ranges = line.split("\\s*;\\s*".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()
+        for (range in ranges) {
+            val osisIds = if (range.indexOf('-') >= 0) {
+                range.split("\\s*-\\s*".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()
+            } else {
+                arrayOf(range)
+            }
+
+            if (osisIds.size == 1) {
+                if (sb.isNotEmpty()) {
+                    sb.append("; ")
+                }
+
+                val osisId = osisIds[0]
+                val readable = osisIdToReadable(line, osisId, null, null)
+                if (readable != null) {
+                    appendScriptureReferenceLink(sb, protocol, osisId, readable)
+                }
+            } else if (osisIds.size == 2) {
+                if (sb.isNotEmpty()) {
+                    sb.append("; ")
+                }
+
+                val bcv = intArrayOf(-1, 0, 0)
+
+                val osisId0 = osisIds[0]
+                val readable0 = osisIdToReadable(line, osisId0, null, bcv)
+                val osisId1 = osisIds[1]
+                val readable1 = osisIdToReadable(line, osisId1, bcv, null)
+                if (readable0 != null && readable1 != null) {
+                    appendScriptureReferenceLink(sb, protocol, "$osisId0-$osisId1", "$readable0-$readable1")
+                }
+            }
+        }
+
+        return sb.toString()
+    }
+
+    private fun appendScriptureReferenceLink(sb: StringBuilder, protocol: String?, osisId: String, readable: String) {
+        if (protocol != null) {
+            sb.append("<a href='")
+            sb.append(protocol)
+            sb.append(':')
+            sb.append(osisId)
+            sb.append("'>")
+        }
+        sb.append(readable)
+        if (protocol != null) {
+            sb.append("</a>")
+        }
+    }
+
+    /**
+     * @param compareWithRangeStart if this is the second part of a range, set this to non-null, with [0] is bookId and [1] chapter_1.
+     * @param outBcv if not null and length is >= 3, will be filled with parsed bcv
+     */
+    private fun osisIdToReadable(line: String, osisId: String, compareWithRangeStart: IntArray?, outBcv: IntArray?): String? {
+        var res: String? = null
+
+        val parts = osisId.split("\\.".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()
+        if (parts.size != 2 && parts.size != 3) {
+            AppLog.w(TAG, "osisId invalid: $osisId in $line")
+        } else {
+            val bookName = parts[0]
+            val chapter_1 = Integer.parseInt(parts[1])
+            val verse_1 = if (parts.size < 3) 0 else Integer.parseInt(parts[2])
+
+            val bookId = OsisBookNames.osisBookNameToBookId(bookName)
+
+            if (outBcv != null && outBcv.size >= 3) {
+                outBcv[0] = bookId
+                outBcv[1] = chapter_1
+                outBcv[2] = verse_1
+            }
+
+            if (bookId < 0) {
+                AppLog.w(TAG, "osisBookName invalid: $bookName in $line")
+            } else {
+                val book = App.services.versions.activeVersion().getBook(bookId)
+
+                if (book != null) {
+                    var full = true
+                    if (compareWithRangeStart != null) {
+                        if (compareWithRangeStart[0] == bookId) {
+                            if (compareWithRangeStart[1] == chapter_1) {
+                                res = verse_1.toString()
+                                full = false
+                            } else {
+                                res = "$chapter_1:$verse_1"
+                                full = false
+                            }
+                        }
+                    }
+
+                    if (full) {
+                        res = if (verse_1 == 0) book.reference(chapter_1) else book.reference(chapter_1, verse_1)
+                    }
+                }
+            }
+        }
+        return res
+    }
+}
